MI-94 moving generator logic to a new file so that it can run seperately from the rest of the generator
